pyproject.toml
[project]
name = "awesome-python-project"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.14.1"
dependencies = [
    "django",
    "djangorestframework",
]

[project.license]
file = "LICENSE"

[project.optional-dependencies]
test = [
    "covdefaults",
    "django-stubs[compatible-mypy]",
    "mypy",
    "pytest",
    "pytest-cov",
    "pytest-django",
    "pytest-randomly",
    "pytest-xdist",
    "time-machine",
]

[build-system]
requires = [
    "setuptools",
]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]

[tool.django-stubs]
django_settings_module = "amphora.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "amphora.settings"

[tool.autoflake]
recursive = true

[tool.mypy]
files = "."
ignore_missing_imports = true
strict = true
exclude = [
    "tests/",
]
plugins = [
    "mypy_django_plugin.main",
]

[tool.coverage.run]
plugins = [
    "covdefaults",
]
omit = [
    "wsgi.py",
    "asgi.py",
    "manage.py",
]

[tool.coverage.html]
directory = "artifacts/htmlcov/"
